Complete fix post status
Fix authorize for post status
