+++ Eravol.UIClient/Pages/PostStatuses/Admin/Index.cshtml
@@ -99,7 +99,7 @@
                                             <td>@item.PostStatusDesc</td>
                                             <td>
                                                 <div class="wt-actionbtn">
-                                                    <a asp-page="/PostStatuses/Clients/Update" asp-route-PostStatusId="@item.PostStatusId" class="wt-addinfo wt-skillsaddinfo"><i class="lnr lnr-pencil"></i></a>
+                                                    <a asp-page="/PostStatuses/Admin/Update" asp-route-PostStatusId="@item.PostStatusId" class="wt-addinfo wt-skillsaddinfo"><i class="lnr lnr-pencil"></i></a>
                                                     <a href="javascript:void(0)" class="wt-deleteinfo" onclick="confirmDeltePostStatus(@item.PostStatusId)"><i class="lnr lnr-trash"></i></a>
                                                 </div>
                                             </td>
